# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "HealthController_alive",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/readyz": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/api/v1/recordings": {
      "post": {
        "description": "Makes the recording and returns the link to PUT the file to.",
        "operationId": "RecordingsController_requestUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start an upload",
        "tags": [
          "recordings"
        ]
      },
      "get": {
        "description": "Newest first.",
        "operationId": "RecordingsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "uploading",
                "uploaded",
                "ready",
                "failed",
                "queued",
                "transcribing",
                "done"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "description": "The id of the last recording of the previous page.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "first",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List recordings",
        "tags": [
          "recordings"
        ]
      }
    },
    "/api/v1/recordings/{id}": {
      "get": {
        "operationId": "RecordingsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One recording",
        "tags": [
          "recordings"
        ]
      },
      "delete": {
        "operationId": "RecordingsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a recording, its audio and its jobs",
        "tags": [
          "recordings"
        ]
      }
    },
    "/api/v1/recordings/{id}/audio": {
      "get": {
        "operationId": "RecordingsController_audio",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A short-lived link to the audio a browser plays",
        "tags": [
          "recordings"
        ]
      }
    },
    "/api/v1/recordings/{id}/transcript": {
      "get": {
        "description": "Every line with both layers: as spoken, and Hinglish.",
        "operationId": "RecordingsController_transcript",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The newest transcript",
        "tags": [
          "recordings"
        ]
      }
    },
    "/api/v1/recordings/{id}/jobs": {
      "post": {
        "operationId": "RecordingsController_createJob",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Queue a transcription",
        "tags": [
          "recordings"
        ]
      },
      "get": {
        "operationId": "RecordingsController_jobs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The jobs of a recording, newest first",
        "tags": [
          "recordings"
        ]
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "operationId": "JobsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{id}/cancel": {
      "post": {
        "operationId": "JobsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stop a waiting or running job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/search": {
      "get": {
        "description": "A few words in either layer, typos allowed; the matches are inside <mark>…</mark>.",
        "operationId": "SearchController_find",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "A few words, in either layer; typos allowed.",
            "schema": {
              "example": "order confirm",
              "type": "string"
            }
          },
          {
            "name": "language",
            "required": false,
            "in": "query",
            "description": "A detected language (ISO 639-1).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "recordingId",
            "required": false,
            "in": "query",
            "description": "Only this recording.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "required": false,
            "in": "query",
            "description": "Transcripts created from this moment (ISO 8601).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "until",
            "required": false,
            "in": "query",
            "description": "Transcripts created up to this moment (ISO 8601).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search every transcript line",
        "tags": [
          "search"
        ]
      }
    },
    "/api/v1/imports": {
      "post": {
        "description": "The connector fetches the call; poll the import, or the recordings, to see it arrive.",
        "operationId": "ImportsController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestImportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch a call from the dialer by its id",
        "tags": [
          "imports"
        ]
      },
      "get": {
        "description": "Newest first.",
        "operationId": "ImportsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "requested",
                "completed",
                "failed"
              ]
            }
          },
          {
            "name": "after",
            "required": false,
            "in": "query",
            "description": "The id of the last import of the previous page.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "first",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List imports",
        "tags": [
          "imports"
        ]
      }
    },
    "/api/v1/imports/{id}": {
      "get": {
        "operationId": "ImportsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One import",
        "tags": [
          "imports"
        ]
      }
    },
    "/events/jobs/{id}": {
      "get": {
        "operationId": "LiveController_job",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Live"
        ]
      }
    },
    "/events/recordings": {
      "get": {
        "operationId": "LiveController_workspace",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Live"
        ]
      }
    }
  },
  "info": {
    "title": "Likho API",
    "description": "Recordings, jobs and transcripts for scripts and connectors. Browsers use /graphql.",
    "version": "1",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "The local gateway"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "An API key from the settings page (lk_...)"
      }
    },
    "schemas": {
      "RequestUploadDto": {
        "type": "object",
        "properties": {
          "originalName": {
            "type": "string",
            "example": "call-2026-10-01-0912.mp3"
          },
          "sizeBytes": {
            "type": "number",
            "example": 105000
          },
          "contentType": {
            "type": "string",
            "example": "audio/mpeg"
          },
          "sha256": {
            "type": "string",
            "description": "When known, the same content is not uploaded twice."
          },
          "externalId": {
            "type": "string",
            "description": "Your own id for the call."
          },
          "source": {
            "type": "string",
            "description": "Where the call comes from: a connector’s name, e.g. ameyo.",
            "example": "ameyo"
          },
          "attributes": {
            "type": "object",
            "description": "Facts about the call: campaign, agent, disposition, call time, ...",
            "example": {
              "campaign": "inbound",
              "agent": "agent-12",
              "disposition": "sale"
            }
          }
        },
        "required": [
          "originalName",
          "sizeBytes"
        ]
      },
      "CreateJobDto": {
        "type": "object",
        "properties": {
          "modelRegistryId": {
            "type": "string",
            "description": "Empty = the default model."
          },
          "languagePolicy": {
            "type": "string",
            "description": "\"auto\" or a language code.",
            "example": "auto"
          },
          "force": {
            "type": "boolean"
          }
        }
      },
      "RequestImportDto": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string",
            "description": "The call’s id in the dialer (its crt_object_id).",
            "example": "d000-0a1b2c3d-vce-0001"
          },
          "source": {
            "type": "string",
            "description": "Which connector; empty = the default one."
          },
          "transcribe": {
            "type": "boolean",
            "description": "Transcribe once stored (default true)."
          }
        },
        "required": [
          "externalId"
        ]
      }
    }
  }
}